+++ roguelike_project/editor/data/buildings_data.json
@@ -14,13 +14,13 @@
         ]
     },
     {
-        "x": 2136,
-        "y": -922,
+        "x": 1998,
+        "y": 1328,
         "image_path": "assets/buildings/temples/catholic.png",
         "solid": true,
         "scale": [
-            980,
-            1470
+            635,
+            951
         ],
         "original_scale": [
             1024,
@@ -56,8 +56,8 @@
         ]
     },
     {
-        "x": 2685,
-        "y": 440,
+        "x": 2193,
+        "y": 166,
         "image_path": "assets/buildings/others/portal.png",
         "solid": true,
         "scale": [
@@ -98,8 +98,8 @@
         ]
     },
     {
-        "x": 1319,
-        "y": 229,
+        "x": 2741,
+        "y": 1299,
         "image_path": "assets/buildings/shops/healer_1.png",
         "solid": true,
         "scale": [
@@ -112,8 +112,8 @@
         ]
     },
     {
-        "x": 2066,
-        "y": 570,
+        "x": 1530,
+        "y": 588,
         "image_path": "assets/buildings/shops/healer.png",
         "solid": true,
         "scale": [
@@ -140,8 +140,8 @@
         ]
     },
     {
-        "x": 364,
-        "y": 1030,
+        "x": 482,
+        "y": 1238,
         "image_path": "assets/buildings/shops/blacksmith.png",
         "solid": true,
         "scale": [
@@ -154,8 +154,8 @@
         ]
     },
     {
-        "x": 4670,
-        "y": 1416,
+        "x": 4522,
+        "y": 1444,
         "image_path": "assets/buildings/houses/orden_house_2.png",
         "solid": true,
         "scale": [
@@ -168,8 +168,8 @@
         ]
     },
     {
-        "x": 1562,
-        "y": 686,
+        "x": 2682,
+        "y": 156,
         "image_path": "assets/buildings/shops/jewlery_shop.png",
         "solid": true,
         "scale": [
@@ -199,34 +199,6 @@
         "x": 3090,
         "y": 848,
         "image_path": "assets/buildings/others/coliseo_2.png",
-        "solid": true,
-        "scale": [
-            1550,
-            1550
-        ],
-        "original_scale": [
-            1024,
-            1024
-        ]
-    },
-    {
-        "x": 3090,
-        "y": 848,
-        "image_path": "assets/buildings/gardens/garden_1.png",
-        "solid": true,
-        "scale": [
-            1550,
-            1550
-        ],
-        "original_scale": [
-            1024,
-            1024
-        ]
-    },
-    {
-        "x": 3090,
-        "y": 848,
-        "image_path": "assets/buildings/gardens/garden_2.png",
         "solid": true,
         "scale": [
             1550,
