+++ roguelike_project/loaders/load_buildings.py
@@ -3,7 +3,7 @@
 import os
 from roguelike_project.entities.buildings.building import Building
 from roguelike_project.systems.editor.json_handler import  save_buildings_to_json
-from roguelike_project.systems.editor.load_buildings_from_json import load_buildings_from_json
+from roguelike_project.systems.editor.buildings.load_buildings_from_json import load_buildings_from_json
 
 BUILDINGS_JSON_PATH = "roguelike_project/editor/data/buildings_data.json"
