+++ roguelike_project/systems/editor/buildings/tools/delete_tool.py
